+++ packages/ui-core/src/components/data-access/Roles/List.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import ModelList from '../ModelList';
+import config from './config';
 
 const List: React.FC = () => {
-    return <ModelList modelName="Role" />;
+    return <ModelList modelName={config.name} />;
 };
 
 export default List;
