+++ packages/ui-core/src/components/data-access/Roles/routes.ts
@@ -1,0 +1,30 @@
+import { PluginRoute } from '@terascope/ui-components';
+import config from './config';
+import Form from './Form';
+import List from './List';
+
+const routes: PluginRoute[] = [
+    {
+        name: `${config.pluralLabel}`,
+        path: `/${config.pathname}`,
+        icon: 'key',
+        component: List,
+        actions: [`/${config.pathname}/create`],
+    },
+    {
+        name: `Create ${config.singularLabel}`,
+        path: `/${config.pathname}/create`,
+        icon: 'add',
+        hidden: true,
+        component: Form,
+    },
+    {
+        name: `Edit ${config.singularLabel}`,
+        path: `/${config.pathname}/edit/:id`,
+        icon: 'pencil alternate',
+        hidden: true,
+        component: Form,
+    },
+];
+
+export default routes;
