+++ packages/ui-core/src/components/data-access/Users/List.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
+import config from './config';
 import ModelList from '../ModelList';
 
 const List: React.FC = () => {
-    return <ModelList modelName="User" />;
+    return <ModelList modelName={config.name} />;
 };
 
 export default List;
